fix proto package renaming :ambulance:
