+++ .gitignore
@@ -20,3 +20,4 @@ project/metals.sbt
 .env
 .DS_Store
 *.iml
+.protogen/
