--- .protogen/src/main/protobuf/lagompb/v1/options.proto
@@ -1,13 +1,0 @@
-syntax = "proto3";
-
-package lagompb.v1;
-
-option java_package = "io.superflat.lagompb.v1.protobuf";
-
-import "scalapb/scalapb.proto";
-
-
-option (scalapb.options) = {
-  scope: PACKAGE
-  preserve_unknown_fields: false
-};
