+++ core/lagompb-core/src/main/protobuf/lagompb/v1/tests.proto
@@ -2,7 +2,7 @@ syntax = "proto3";
 
 package lagompb.v1;
 
-option java_package = "io.superflat.lagompb.v1.protobuf";
+option java_package = "io.superflat.lagompb.protobuf.v1";
 
 import "submodules/protobuf/lagompb/v1/extensions.proto";
 import "google/protobuf/any.proto";
@@ -43,12 +43,12 @@ message TestFailCmd {
 }
 
 // Used to test event handler failure in the aggregate root
-message TestEventFailureCmd{
+message TestEventFailureCmd {
     string company_uuid = 1 [(lagompb.v1.command).entity_id = true];
 }
 
 // Event emitted to test event handler failure in the aggregate root
-message TestEventFailure{
+message TestEventFailure {
 
 }
