+++ core/lagompb-core/src/main/resources/lagompb.conf
@@ -401,7 +401,7 @@ lagompb {
 
 	# Protocol buffers package name to allow lagompb to
 	# automatically build a descriptor registry
-	protos-package = "io.superflat.lagompb.v1.protobuf"
+	protos-package = "io.superflat.lagompb.protobuf.v1"
 	protos-package = ${?PROTO_PACKAGE}
 
 	# Ask timeout is required to
