+++ core/lagompb-core/src/main/scala/io/superflat/lagompb/AggregateRoot.scala
@@ -9,13 +9,13 @@ import akka.persistence.typed.PersistenceId
 import akka.persistence.typed.scaladsl.{Effect, EventSourcedBehavior, ReplyEffect, RetentionCriteria}
 import com.google.protobuf.any.Any
 import io.superflat.lagompb.encryption.EncryptionAdapter
-import io.superflat.lagompb.v1.protobuf.core._
-import io.superflat.lagompb.v1.protobuf.core.CommandHandlerResponse.HandlerResponse.{
+import io.superflat.lagompb.protobuf.v1.core._
+import io.superflat.lagompb.protobuf.v1.core.CommandHandlerResponse.HandlerResponse.{
   Empty,
   FailedResponse,
   SuccessResponse
 }
-import io.superflat.lagompb.v1.protobuf.core.SuccessCommandHandlerResponse.Response.{Event, NoEvent}
+import io.superflat.lagompb.protobuf.v1.core.SuccessCommandHandlerResponse.Response.{Event, NoEvent}
 import org.slf4j.{Logger, LoggerFactory}
 import scalapb.{GeneratedMessage, GeneratedMessageCompanion}
