+++ core/lagompb-core/src/main/scala/io/superflat/lagompb/encryption/NoEncryption.scala
@@ -2,7 +2,7 @@ package io.superflat.lagompb.encryption
 
 import com.google.protobuf.ByteString
 import com.google.protobuf.any.Any
-import io.superflat.lagompb.v1.protobuf.encryption.EncryptedProto
+import io.superflat.lagompb.protobuf.v1.encryption.EncryptedProto
 
 import scala.util.Try
