+++ core/lagompb-core/src/test/scala/io/superflat/lagompb/AggregateRootSpec.scala
@@ -12,9 +12,9 @@ import com.google.protobuf.any.Any
 import io.superflat.lagompb.data.{TestAggregateRoot, TestCommandHandler, TestEventHandler}
 import io.superflat.lagompb.encryption.EncryptionAdapter
 import io.superflat.lagompb.testkit.BaseActorTestKit
-import io.superflat.lagompb.v1.protobuf.core._
-import io.superflat.lagompb.v1.protobuf.core.CommandReply.Reply
-import io.superflat.lagompb.v1.protobuf.tests._
+import io.superflat.lagompb.protobuf.v1.core._
+import io.superflat.lagompb.protobuf.v1.core.CommandReply.Reply
+import io.superflat.lagompb.protobuf.v1.tests._
 
 import scala.concurrent.duration.FiniteDuration
 
@@ -67,7 +67,7 @@ class AggregateRootSpec extends BaseActorTestKit(s"""
         Map("audit|employeeUuid" -> "1223", "audit|createdAt" -> "2020-04-17")
       aggregateRef ! Command(testCmd, commandSender.ref, data)
       commandSender.receiveMessage(replyTimeout) match {
-        case CommandReply(reply) =>
+        case CommandReply(reply, _) =>
           reply match {
             case Reply.Empty => fail("unexpected message state")
             case Reply.SuccessfulReply(value) =>
@@ -103,7 +103,7 @@ class AggregateRootSpec extends BaseActorTestKit(s"""
       aggregateRef ! Command(testCmd, commandSender.ref, Map.empty[String, String])
 
       commandSender.receiveMessage(replyTimeout) match {
-        case CommandReply(reply) =>
+        case CommandReply(reply, _) =>
           reply match {
             case Reply.FailedReply(value) =>
               value.reason shouldBe "command is invalid"
@@ -137,7 +137,7 @@ class AggregateRootSpec extends BaseActorTestKit(s"""
         Map("audit|employeeUuid" -> "1223", "audit|createdAt" -> "2020-04-17")
       aggregateRef ! Command(testCmd, commandSender.ref, data)
       commandSender.receiveMessage(replyTimeout) match {
-        case CommandReply(reply) =>
+        case CommandReply(reply, _) =>
           reply match {
             case Reply.Empty => fail("unexpected message state")
             case Reply.SuccessfulReply(value) =>
@@ -172,7 +172,7 @@ class AggregateRootSpec extends BaseActorTestKit(s"""
       // let us send the command to the aggregate
       aggregateRef ! Command(testCmd, commandSender.ref, Map.empty)
       commandSender.receiveMessage(replyTimeout) match {
-        case CommandReply(reply) =>
+        case CommandReply(reply, _) =>
           reply match {
             case Reply.FailedReply(value) =>
               value.reason should include("unknown command handler response")
@@ -204,7 +204,7 @@ class AggregateRootSpec extends BaseActorTestKit(s"""
       // let us send the command to the aggregate
       aggregateRef ! Command(testCmd, commandSender.ref, Map.empty)
       commandSender.receiveMessage(replyTimeout) match {
-        case CommandReply(reply) =>
+        case CommandReply(reply, _) =>
           reply match {
             case Reply.FailedReply(value) =>
               value.reason should include("unknown command handler success response")
@@ -236,7 +236,7 @@ class AggregateRootSpec extends BaseActorTestKit(s"""
       // let us send the command to the aggregate
       aggregateRef ! Command(testCmd, commandSender.ref, Map.empty)
       commandSender.receiveMessage(replyTimeout) match {
-        case CommandReply(reply) =>
+        case CommandReply(reply, _) =>
           reply match {
             case Reply.FailedReply(value) =>
               value.reason should include("[Lagompb] unable to parse event")
@@ -364,7 +364,7 @@ class AggregateRootSpec extends BaseActorTestKit(s"""
       // let us send the command to the aggregate
       aggregateRef ! Command(testCmd, commandSender.ref, Map.empty)
       commandSender.receiveMessage(replyTimeout) match {
-        case CommandReply(reply) =>
+        case CommandReply(reply, _) =>
           reply match {
             case Reply.FailedReply(value) =>
               value.reason should include("an implementation is missing")
