+++ core/lagompb-core/src/test/scala/io/superflat/lagompb/BaseServiceImplSpec.scala
@@ -8,9 +8,9 @@ import com.opentable.db.postgres.embedded.EmbeddedPostgres
 import io.superflat.lagompb.data._
 import io.superflat.lagompb.encryption.EncryptionAdapter
 import io.superflat.lagompb.testkit.BaseSpec
-import io.superflat.lagompb.v1.protobuf.core._
-import io.superflat.lagompb.v1.protobuf.core.CommandReply.Reply
-import io.superflat.lagompb.v1.protobuf.tests.{TestCmd, TestState}
+import io.superflat.lagompb.protobuf.v1.core._
+import io.superflat.lagompb.protobuf.v1.core.CommandReply.Reply
+import io.superflat.lagompb.protobuf.v1.tests.{TestCmd, TestState}
 
 import scala.concurrent.ExecutionContext.Implicits.global
