+++ core/lagompb-core/src/test/scala/io/superflat/lagompb/ProtosExtensionSpec.scala
@@ -3,8 +3,8 @@ package io.superflat.lagompb
 import java.util.UUID
 
 import io.superflat.lagompb.testkit.BaseSpec
-import io.superflat.lagompb.v1.protobuf.extensions.ExtensionsProto
-import io.superflat.lagompb.v1.protobuf.tests.{TestCmd, TestEvent}
+import io.superflat.lagompb.protobuf.v1.extensions.ExtensionsProto
+import io.superflat.lagompb.protobuf.v1.tests.{TestCmd, TestEvent}
 import scalapb.descriptors.FieldDescriptor
 
 class ProtosExtensionSpec extends BaseSpec {
