+++ core/lagompb-core/src/test/scala/io/superflat/lagompb/ProtosRegistrySpec.scala
@@ -2,9 +2,9 @@ package io.superflat.lagompb
 
 import com.google.protobuf.any.Any
 import io.superflat.lagompb.testkit.BaseSpec
-import io.superflat.lagompb.v1.protobuf.core.CoreProto
-import io.superflat.lagompb.v1.protobuf.options.OptionsProto
-import io.superflat.lagompb.v1.protobuf.tests.TestCmd
+import io.superflat.lagompb.protobuf.v1.core.CoreProto
+
+import io.superflat.lagompb.protobuf.v1.tests.TestCmd
 import scalapb.{GeneratedMessage, GeneratedMessageCompanion}
 
 class ProtosRegistrySpec extends BaseSpec {
@@ -14,7 +14,6 @@ class ProtosRegistrySpec extends BaseSpec {
     "succeed" in {
       val fos = ProtosRegistry.registry
       fos.size should be >= 1
-      fos should contain(OptionsProto)
       fos should contain(CoreProto)
     }
