+++ core/lagompb-core/src/test/scala/io/superflat/lagompb/data/TestEventHandler.scala
@@ -2,17 +2,17 @@ package io.superflat.lagompb.data
 
 import akka.actor.ActorSystem
 import io.superflat.lagompb.EventHandler
-import io.superflat.lagompb.v1.protobuf.core.MetaData
-import io.superflat.lagompb.v1.protobuf.tests.{TestEvent, TestEventFailure, TestState}
+import io.superflat.lagompb.protobuf.v1.core.MetaData
+import io.superflat.lagompb.protobuf.v1.tests.{TestEvent, TestEventFailure, TestState}
 
 class TestEventHandler(actorSystem: ActorSystem) extends EventHandler[TestState](actorSystem) {
 
   override def handle(event: scalapb.GeneratedMessage, currentState: TestState, eventMeta: MetaData): TestState =
     event match {
-      case TestEvent(companyUuid, name) =>
+      case TestEvent(companyUuid, name, _) =>
         handleTestEvent(companyUuid, name, currentState)
-      case TestEventFailure() => throw new NotImplementedError()
-      case _                  => throw new NotImplementedError()
+      case TestEventFailure(_) => throw new NotImplementedError()
+      case _                   => throw new NotImplementedError()
     }
 
   private[this] def handleTestEvent(companyUuid: String, name: String, state: TestState): TestState =
