+++ core/lagompb-readside/src/main/scala/io/superflat/lagompb/readside/ReadSideEvent.scala
@@ -1,6 +1,6 @@
 package io.superflat.lagompb.readside
 
-import io.superflat.lagompb.v1.protobuf.core.MetaData
+import io.superflat.lagompb.protobuf.v1.core.MetaData
 
 /**
  * Wraps events read from the journal and make it available to any consumer
