+++ project/CoverageWhitelist.scala
@@ -2,7 +2,7 @@ object CoverageWhitelist {
 
   val whitelist = Seq(
     "<empty>",
-    "io.superflat.lagompb.v1.protobuf.*",
+    "io.superflat.lagompb.protobuf.*", // ignore scalapb generated classes
     "io.superflat.lagompb.data.*",
     "io.superflat.lagompb.testkit.*",
     "io.superflat.lagompb.readside.*",
